+++ framework/directives/splitView.js
@@ -93,12 +93,10 @@
           $onsen.declareVarAttribute(attrs, splitView);
 
           element.data('ons-split-view', splitView);
-          $onsen.aliasStack.register('ons.splitView', splitView);
 
           scope.$on('$destroy', function() {
             splitView._events = undefined;
             element.data('ons-split-view', undefined);
-            $onsen.aliasStack.unregister('ons.splitView', splitView);
           });
 
           $onsen.fireComponentEvent(element[0], 'init');